chainId in SequencePay is now int64

+++ Plugins/SequencePlugin/Source/SequencePlugin/Public/Marketplace/Structs/SeqGetSwapPriceArgs.h
@@ -25,7 +25,7 @@ public:
     FString BuyAmount;
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "0xSequence")
-    int ChainID;
+    int64 ChainID;
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "0xSequence")
     int SlippagePercentage;
